requirements.txt
APScheduler==3.10.1
asgiref==3.5.2
asttokens==2.2.1
beautifulsoup4==4.11.2
channels==4.0.0
click==8.1.3
colorama==0.4.6
crispy-bootstrap5==0.7
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==2.0.0
Django==4.1.4
django-apscheduler==0.6.2
django-crispy-forms==1.14.0
django-crontab==0.7.1
django-dbbackup==4.0.2
django-filter==22.1
django-import-export==3.0.2
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-bulk==0.2.1
et-xmlfile==1.1.0
executing==1.2.0
future==0.18.3
gunicorn==20.1.0
i==2.1.0
Markdown==3.4.1
MarkupPy==1.14
numpy==1.24.1
odfpy==1.4.1
openpyxl==3.0.10
pandas==1.5.3
Pillow==9.5.0
plan==0.5
pyfingerprint==1.5
Pygments==2.15.0
pyserial==3.5
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzk==0.9
schedule==1.1.0
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tablib==3.3.0
tornado==6.2
typing_extensions==4.6.0
tzdata==2022.7
tzlocal==4.2
whitenoise==6.4.0
xlrd==2.0.1
XlsxWriter==3.0.7
xlwt==1.3.0
